CT3/requirements.txt
accelerate
activeft
datasets
docarray
faiss-cpu
fastapi
lm-eval==0.4.7
numpy
peft
protobuf
pydantic
sentencepiece
setuptools
torch==2.5.1
transformers
uvicorn
wheel
git+https://github.com/bigcode-project/bigcode-evaluation-harness.git@6116c6a#egg=bigcode-eval
